refactor: reorganize three.js components
