+++ app/page.tsx
@@ -1,10 +1,14 @@
 "use client";
 
+import dynamic from "next/dynamic";
 import Link from "next/link";
-import Experience from "../components/Experience";
 import Navbar from "../components/Navbar";
 import { useTranslation } from "react-i18next";
 import "./i18n/config";
+
+const Experience = dynamic(() => import("../components/three/Experience"), {
+  ssr: false,
+});
 
 export default function HomePage() {
   const { t } = useTranslation();
