+++ app/work/page.tsx
@@ -2,10 +2,15 @@
 
 import { useState } from "react";
 import Image from "next/image";
-import Experience from "../../components/Experience";
+import dynamic from "next/dynamic";
 import Navbar from "../../components/Navbar";
 import { useTranslation } from "react-i18next";
 import "../i18n/config";
+
+const Experience = dynamic(
+  () => import("../../components/three/Experience"),
+  { ssr: false }
+);
 
 const projectOrder = ["aurora", "mare", "spectrum"] as const;
